Minor code refactoring
