+++ Microsoft.Toolkit.Mvvm/Messaging/Internals/System/Gen2GcCallback.cs
@@ -1,5 +1,6 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
 
 using System.Runtime.ConstrainedExecution;
 using System.Runtime.InteropServices;
